Changes and corrections
