+++ .travis.yml
@@ -1,5 +1,5 @@
 dist: trusty
-sudo: false
+sudo: require
 
 language: python
 python:
@@ -21,9 +21,17 @@ env:
   - TEST_PLATFORM="LPC1768"
   - TEST_PLATFORM="STM32F1"
   - TEST_PLATFORM="teensy35"
-  - TEST_PLATFORM="linux_x86_64"
+  - TEST_PLATFORM="linux_native"
+
+addons:
+  apt:
+    sources:
+      - ubuntu-toolchain-r-test
+    packages:
+      - g++-7
 
 before_install:
+  - sudo update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-7 90
   #
   # Fetch the tag information for the current branch
   - git fetch origin --tags
