+++ Marlin/src/HAL/HAL_LINUX/HAL_timers.cpp
@@ -20,7 +20,7 @@
  *
  */
 
-#ifdef __PLAT_X86_64__
+#ifdef __PLAT_LINUX__
 
 #include "hardware/Timer.h"
 
@@ -38,8 +38,8 @@ HAL_TEMP_TIMER_ISR;
 Timer timers[2];
 
 void HAL_timer_init(void) {
-  timers[0].init(0, HAL_STEPPER_TIMER_RATE, TIMER0_IRQHandler);
-  timers[1].init(1, HAL_TEMP_TIMER_RATE, TIMER1_IRQHandler);
+  timers[0].init(0, STEPPER_TIMER_RATE, TIMER0_IRQHandler);
+  timers[1].init(1, TEMP_TIMER_RATE, TIMER1_IRQHandler);
 }
 
 void HAL_timer_start(const uint8_t timer_num, const uint32_t frequency) {
@@ -70,4 +70,4 @@ hal_timer_t HAL_timer_get_count(const uint8_t timer_num) {
   return timers[timer_num].getCount();
 }
 
-#endif // __PLAT_X86_64__
+#endif // __PLAT_LINUX__
