+++ Marlin/src/HAL/HAL_LINUX/hardware/Gpio.cpp
@@ -20,31 +20,11 @@
  *
  */
 
-#ifdef __PLAT_X86_64__
+#ifdef __PLAT_LINUX__
 
-#include "IOLoggerCSV.h"
+#include "Gpio.h"
 
-IOLoggerCSV::IOLoggerCSV(std::string filename) {
-  file.open(filename);
-}
+pin_data Gpio::pin_map[Gpio::pin_count+1] = {};
+IOLogger* Gpio::logger = nullptr;
 
-IOLoggerCSV::~IOLoggerCSV() {
-  file.close();
-}
-
-void IOLoggerCSV::log(GpioEvent ev) {
-  std::lock_guard<std::mutex> lock(vector_lock);
-  events.push_back(ev); //minimal impact to signal handler
-}
-
-void IOLoggerCSV::flush() {
-  { std::lock_guard<std::mutex> lock(vector_lock);
-    while (!events.empty()) {
-      file << events.front().timestamp << ", "<< events.front().pin_id << ", " << events.front().event << std::endl;
-      events.pop_front();
-    }
-  }
-  file.flush();
-}
-
-#endif // __PLAT_X86_64__
+#endif // __PLAT_LINUX__
