+++ Marlin/src/HAL/HAL_LINUX/hardware/LinearAxis.cpp
@@ -20,7 +20,7 @@
  *
  */
 
-#ifdef __PLAT_X86_64__
+#ifdef __PLAT_LINUX__
 
 #include <random>
 #include <stdio.h>
@@ -63,4 +63,4 @@ void LinearAxis::interrupt(GpioEvent ev) {
   }
 }
 
-#endif // __PLAT_X86_64__
+#endif // __PLAT_LINUX__
