+++ Marlin/src/HAL/HAL_LINUX/include/Arduino.h
@@ -24,6 +24,7 @@
 #include <stddef.h>
 #include <stdint.h>
 #include <math.h>
+#include <cstring>
 
 #include <pinmapping.h>
 
@@ -84,6 +85,7 @@ extern "C" void GpioDisableInt(uint32_t port, uint32_t pin);
 #define pgm_read_word(addr)       pgm_read_word_near(addr)
 #define pgm_read_dword(addr)      pgm_read_dword_near(addr)
 
+using std::memcpy;
 #define memcpy_P memcpy
 #define sprintf_P sprintf
 #define strstr_P strstr
