+++ Marlin/src/HAL/HAL_LINUX/include/pinmapping.cpp
@@ -20,7 +20,7 @@
  *
  */
 
-#ifdef __PLAT_X86_64__
+#ifdef __PLAT_LINUX__
 
 #include <pinmapping.h>
 
@@ -67,4 +67,4 @@ int16_t PARSED_PIN_INDEX(const char code, const int16_t dval) {
   return parser.intval(code, dval);
 }
 
-#endif // __PLAT_X86_64__
+#endif // __PLAT_LINUX__
