+++ Marlin/src/HAL/platforms.h
@@ -41,8 +41,8 @@
   #define HAL_PLATFORM HAL_STM32
 #elif defined(ARDUINO_ARCH_ESP32)
   #define HAL_PLATFORM HAL_ESP32
-#elif defined(__PLAT_X86_64__)
-  #define HAL_PLATFORM HAL_X86_64
+#elif defined(__PLAT_LINUX__)
+  #define HAL_PLATFORM HAL_LINUX
 #else
   #error "Unsupported Platform!"
 #endif
