+++ Marlin/src/pins/pins.h
@@ -449,11 +449,11 @@
   #include "pins_ESP32.h"
 
 //
-// x86 Native Debug board
+// Linux Native Debug board
 //
 
-#elif MB(LINUX_X86_64)
-  #include "pins_RAMPS_X86.h"         // x86 Debug Build
+#elif MB(LINUX_RAMPS)
+  #include "pins_RAMPS_Linux.h"         // Linux Debug Build
 
 #else
   #error "Unknown MOTHERBOARD value set in Configuration.h"
