+++ buildroot/share/tests/linux_native-tests
@@ -7,7 +7,7 @@
 set -e
 
 restore_configs
-opt_set MOTHERBOARD BOARD_RAMPS_X86
+opt_set MOTHERBOARD BOARD_LINUX_RAMPS
 opt_set TEMP_SENSOR_BED 1
 opt_enable PIDTEMPBED EEPROM_SETTINGS
 exec_test $1 $2 "Linux with EEPROM"
