+++ platformio.ini
@@ -386,12 +386,12 @@ monitor_speed     = 250000
 # Native
 # No supported Arduino libraries, base Marlin only
 #
-[env:linux_x86_64]
+[env:linux_native]
 platform        = native
-build_flags     = -D__PLAT_X86_64__ -std=c++11 -ggdb -g -lrt -lpthread
-src_build_flags = -Wall -IMarlin/src/HAL/HAL_X86_64/include
+build_flags     = -D__PLAT_LINUX__ -std=gnu++17 -ggdb -g -lrt -lpthread
+src_build_flags = -Wall -IMarlin/src/HAL/HAL_LINUX/include
 build_unflags   = -Wall
 lib_ldf_mode    = off
 lib_deps        =
 extra_scripts   =
-src_filter      = ${common.default_src_filter}
+src_filter      = ${common.default_src_filter} +<src/HAL/HAL_LINUX>
